Move utils tests back to video manager tests.
